Fixed replace tests
